# Optimize delta decoding (part 1)