--- a/encoding/delta/binary_packed.go
+++ b/encoding/delta/binary_packed.go
@@ -9,6 +9,7 @@ import (
 
 	"github.com/segmentio/parquet-go/encoding"
 	"github.com/segmentio/parquet-go/format"
+	"github.com/segmentio/parquet-go/internal/bitpack"
 	"github.com/segmentio/parquet-go/internal/unsafecast"
 )
 
@@ -112,7 +113,7 @@ func encodeInt32Default(dst []byte, src []int32) []byte {
 		for i, bitWidth := range bitWidths {
 			if bitWidth != 0 {
 				miniBlock := (*[miniBlockSize]int32)(block[i*miniBlockSize:])
-				miniBlockPackInt32(dst[n:], miniBlock, uint(bitWidth))
+				encodeMiniBlockInt32(dst[n:], miniBlock, uint(bitWidth))
 				n += (miniBlockSize * int(bitWidth)) / 8
 			}
 		}
@@ -158,7 +159,7 @@ func encodeInt64Default(dst []byte, src []int64) []byte {
 		for i, bitWidth := range bitWidths {
 			if bitWidth != 0 {
 				miniBlock := (*[miniBlockSize]int64)(block[i*miniBlockSize:])
-				miniBlockPackInt64(dst[n:], miniBlock, uint(bitWidth))
+				encodeMiniBlockInt64(dst[n:], miniBlock, uint(bitWidth))
 				n += (miniBlockSize * int(bitWidth)) / 8
 			}
 		}
@@ -279,74 +280,6 @@ func blockBitWidthsInt64(bitWidths *[numMiniBlocks]byte, block *[blockSize]int64
 	}
 }
 
-func resize(buf []byte, size int) []byte {
-	if cap(buf) < size {
-		return grow(buf, size)
-	}
-	if size > len(buf) {
-		clear := buf[len(buf):size]
-		for i := range clear {
-			clear[i] = 0
-		}
-	}
-	return buf[:size]
-}
-
-func grow(buf []byte, size int) []byte {
-	newCap := 2 * cap(buf)
-	if newCap < size {
-		newCap = size
-	}
-	newBuf := make([]byte, size, newCap)
-	copy(newBuf, buf)
-	return newBuf
-}
-
-var (
-	// These variables are used to hook optimized versions of the functions
-	// decoding delta encoded blocks and mini-blocks.
-	decodeBlockInt32 = decodeBlockInt32Default
-	decodeBlockInt64 = decodeBlockInt64Default
-
-	decodeMiniBlockInt32Table [32]func(dst []int32, src []uint32)
-	decodeMiniBlockInt64Table [64]func(dst []int64, src []uint32)
-)
-
-func init() {
-	for i := range decodeMiniBlockInt32Table {
-		bitWidth := uint(i + 1)
-		decodeMiniBlockInt32Table[i] = func(dst []int32, src []uint32) {
-			decodeMiniBlockInt32Default(dst, src, bitWidth)
-		}
-	}
-
-	for i := range decodeMiniBlockInt64Table {
-		bitWidth := uint(i + 1)
-		decodeMiniBlockInt64Table[i] = func(dst []int64, src []uint32) {
-			decodeMiniBlockInt64Default(dst, src, bitWidth)
-		}
-	}
-
-	decodeMiniBlockInt32Table[31] = decodeMiniBlockInt32x32bits
-	decodeMiniBlockInt64Table[63] = decodeMiniBlockInt64x64bits
-}
-
-func decodeMiniBlockInt32x32bits(dst []int32, src []uint32) {
-	copy(dst, unsafecast.Uint32ToInt32(src))
-}
-
-func decodeMiniBlockInt64x64bits(dst []int64, src []uint32) {
-	copy(dst, unsafecast.Uint32ToInt64(src))
-}
-
-func decodeMiniBlockInt32(dst []int32, src []uint32, bitWidth uint) {
-	decodeMiniBlockInt32Table[bitWidth-1](dst, src)
-}
-
-func decodeMiniBlockInt64(dst []int64, src []uint32, bitWidth uint) {
-	decodeMiniBlockInt64Table[bitWidth-1](dst, src)
-}
-
 func decodeInt32(dst, src []byte) ([]byte, []byte, error) {
 	blockSize, numMiniBlocks, totalValues, firstValue, src, err := decodeBinaryPackedHeader(src)
 	if err != nil {
@@ -368,6 +301,9 @@ func decodeInt32(dst, src []byte) ([]byte, []byte, error) {
 	lastValue := int32(firstValue)
 	numValuesInMiniBlock := blockSize / numMiniBlocks
 
+	const padding = 16
+	miniBlockTemp := make([]byte, 256+padding)
+
 	for totalValues > 0 && len(src) > 0 {
 		var minDelta int64
 		var bitWidths []byte
@@ -384,11 +320,15 @@ func decodeInt32(dst, src []byte) ([]byte, []byte, error) {
 				miniBlockSize := (numValuesInMiniBlock * int(bitWidth)) / 8
 				miniBlockData := src
 				if miniBlockSize <= len(src) {
-					miniBlockData = src[:miniBlockSize]
+					miniBlockData = miniBlockData[:miniBlockSize]
 				}
 				src = src[len(miniBlockData):]
-				in := unsafecast.BytesToUint32(miniBlockData)
-				decodeMiniBlockInt32(out[writeOffset:writeOffset+n], in, uint(bitWidth))
+				if cap(miniBlockData) < miniBlockSize+bitpack.Padding {
+					miniBlockTemp = resize(miniBlockTemp[:0], miniBlockSize+bitpack.Padding)
+					miniBlockData = miniBlockTemp[:copy(miniBlockTemp, miniBlockData)]
+				}
+				miniBlockData = miniBlockData[:miniBlockSize]
+				bitpack.UnpackInt32(out[writeOffset:writeOffset+n], miniBlockData, uint(bitWidth))
 			}
 			writeOffset += n
 			totalValues -= n
@@ -425,6 +365,9 @@ func decodeInt64(dst, src []byte) ([]byte, []byte, error) {
 	lastValue := firstValue
 	numValuesInMiniBlock := blockSize / numMiniBlocks
 
+	const padding = 16
+	miniBlockTemp := make([]byte, 512+padding)
+
 	for totalValues > 0 && len(src) > 0 {
 		var minDelta int64
 		var bitWidths []byte
@@ -443,8 +386,12 @@ func decodeInt64(dst, src []byte) ([]byte, []byte, error) {
 					miniBlockData = src[:miniBlockSize]
 				}
 				src = src[len(miniBlockData):]
-				in := unsafecast.BytesToUint32(miniBlockData)
-				decodeMiniBlockInt64(out[writeOffset:writeOffset+n], in, uint(bitWidth))
+				if len(miniBlockData) < miniBlockSize+bitpack.Padding {
+					miniBlockTemp = resize(miniBlockTemp[:0], miniBlockSize+bitpack.Padding)
+					miniBlockData = miniBlockTemp[:copy(miniBlockTemp, miniBlockData)]
+				}
+				miniBlockData = miniBlockData[:miniBlockSize]
+				bitpack.UnpackInt64(out[writeOffset:writeOffset+n], miniBlockData, uint(bitWidth))
 			}
 			writeOffset += n
 			totalValues -= n