--- a/encoding/delta/binary_packed_amd64.go
+++ b/encoding/delta/binary_packed_amd64.go
@@ -10,6 +10,8 @@ func init() {
 	if cpu.X86.HasAVX2 {
 		encodeInt32 = encodeInt32AVX2
 		encodeInt64 = encodeInt64AVX2
+		decodeBlockInt32 = decodeBlockInt32AVX2
+		decodeBlockInt64 = decodeBlockInt64AVX2
 	}
 }
 
@@ -191,3 +193,9 @@ func encodeInt64AVX2(dst []byte, src []int64) []byte {
 
 	return dst
 }
+
+//go:noescape
+func decodeBlockInt32AVX2(dst []int32, minDelta, lastValue int32) int32
+
+//go:noescape
+func decodeBlockInt64AVX2(dst []int64, minDelta, lastValue int64) int64