--- a/encoding/delta/binary_packed_decode_amd64.s
+++ b/encoding/delta/binary_packed_decode_amd64.s
@@ -1,0 +1,89 @@
+//go:build !purego
+
+#include "textflag.h"
+
+// -----------------------------------------------------------------------------
+// 32 bits
+// -----------------------------------------------------------------------------
+
+// func decodeBlockInt32AVX2(dst []int32, minDelta, lastValue int32) int32
+TEXT ·decodeBlockInt32AVX2(SB), NOSPLIT, $0-36
+    MOVQ dst_base+0(FP), AX
+    MOVQ dst_len+8(FP), BX
+    MOVLQZX minDelta+24(FP), CX
+    MOVLQZX lastValue+28(FP), DX
+    XORQ SI, SI
+
+    CMPQ BX, $8
+    JB test
+
+    MOVQ BX, DI
+    SHRQ $3, DI
+    SHLQ $3, DI
+
+    VPXOR X1, X1, X1
+    MOVQ CX, X0
+    MOVQ DX, X1
+    VPBROADCASTD X0, Y0
+loopAVX2:
+    VMOVDQU (AX)(SI*4), Y2
+    VPADDD Y0, Y2, Y2 // Y2[:] += minDelta
+    VPADDD Y1, Y2, Y2 // Y2[0] += lastValue
+
+    VPSLLDQ $4, Y2, Y3
+    VPADDD Y3, Y2, Y2
+
+    VPSLLDQ $8, Y2, Y3
+    VPADDD Y3, Y2, Y2
+
+    VPSHUFD $0xFF, X2, X1
+    VPERM2I128 $1, Y2, Y2, Y3
+    VPADDD X1, X3, X3
+
+    VMOVDQU X2, (AX)(SI*4)
+    VMOVDQU X3, 16(AX)(SI*4)
+    VPSRLDQ $12, X3, X1 // lastValue
+
+    ADDQ $8, SI
+    CMPQ SI, DI
+    JNE loopAVX2
+    VZEROUPPER
+    MOVQ X1, DX
+    JMP test
+loop:
+    MOVL (AX)(SI*4), DI
+    ADDL CX, DI
+    ADDL DI, DX
+    MOVL DX, (AX)(SI*4)
+    INCQ SI
+test:
+    CMPQ SI, BX
+    JNE loop
+done:
+    MOVL DX, ret+32(FP)
+    RET
+
+// -----------------------------------------------------------------------------
+// 64 bits
+// -----------------------------------------------------------------------------
+
+// func decodeBlockInt64AVX2(dst []int64, minDelta, lastValue int64) int64
+TEXT ·decodeBlockInt64AVX2(SB), NOSPLIT, $0-48
+    MOVQ dst_base+0(FP), AX
+    MOVQ dst_len+8(FP), BX
+    MOVQ minDelta+24(FP), CX
+    MOVQ lastValue+32(FP), DX
+    XORQ SI, SI
+    JMP test
+loop:
+    MOVQ (AX)(SI*8), DI
+    ADDQ CX, DI
+    ADDQ DI, DX
+    MOVQ DX, (AX)(SI*8)
+    INCQ SI
+test:
+    CMPQ SI, BX
+    JNE loop
+done:
+    MOVQ DX, ret+40(FP)
+    RET