This PR contributes to https://github.com/segmentio/parquet-go/issues/226 by optimizing delta decoding.

```
name                                             old time/op   new time/op    delta
Decode/DELTA_BINARY_PACKED/int32                   257µs ± 0%      20µs ± 0%    -92.38%  (p=0.000 n=9+10)
Decode/DELTA_BINARY_PACKED/int64                   257µs ± 0%      24µs ± 0%    -90.81%  (p=0.000 n=9+10)
Decode/DELTA_LENGTH_BYTE_ARRAY/byte_array          287µs ±14%      90µs ± 2%    -68.70%  (p=0.000 n=10+10)
Decode/DELTA_BYTE_ARRAY/byte_array                 470µs ±17%     156µs ± 1%    -66.73%  (p=0.000 n=10+10)
Decode/DELTA_BYTE_ARRAY/fixed_length_byte_array    265µs ±10%      92µs ± 0%    -65.40%  (p=0.000 n=8+9)

name                                             old speed     new speed      delta
Decode/DELTA_BINARY_PACKED/int32                 156MB/s ± 0%  2043MB/s ± 0%  +1211.86%  (p=0.000 n=9+10)
Decode/DELTA_BINARY_PACKED/int64                 311MB/s ± 0%  3387MB/s ± 0%   +988.35%  (p=0.000 n=9+10)
Decode/DELTA_LENGTH_BYTE_ARRAY/byte_array        506MB/s ±13%  1613MB/s ± 2%   +218.53%  (p=0.000 n=10+10)
Decode/DELTA_BYTE_ARRAY/byte_array               310MB/s ±15%   925MB/s ± 1%   +198.66%  (p=0.000 n=10+10)
Decode/DELTA_BYTE_ARRAY/fixed_length_byte_array  588MB/s ±24%  1747MB/s ± 0%   +197.23%  (p=0.000 n=9+9)

name                                             old value/s   new value/s    delta
Decode/DELTA_BINARY_PACKED/int32                   38.9M ± 0%    510.7M ± 0%  +1211.87%  (p=0.000 n=9+10)
Decode/DELTA_BINARY_PACKED/int64                   38.9M ± 0%    423.5M ± 0%   +988.36%  (p=0.000 n=9+10)
Decode/DELTA_LENGTH_BYTE_ARRAY/byte_array          35.0M ±13%    111.6M ± 2%   +218.53%  (p=0.000 n=10+10)
Decode/DELTA_BYTE_ARRAY/byte_array                 21.4M ±15%     64.0M ± 1%   +198.67%  (p=0.000 n=10+10)
Decode/DELTA_BYTE_ARRAY/fixed_length_byte_array    36.7M ±24%    109.2M ± 0%   +197.24%  (p=0.000 n=9+9)
```

Here is an example of the impact it has on the merge benchmarks:
```
name                                 old time/op  new time/op  delta
MergeFiles/UUID/groups=3,rows=30000  42.0µs ± 1%  20.0µs ± 4%   -52.47%  (p=0.000 n=10+10)

name                                 old row/s    new row/s    delta
MergeFiles/UUID/groups=3,rows=30000   23.4M ± 1%   49.3M ± 4%  +110.47%  (p=0.000 n=10+10)
```

There is more room for improvement by adding specializations for decoding bit-packed values of various bit widths.